Modified filter_single_record to accept list of accepted target values
This should flow cleanly through from the config file JSON to allow the testing of lists of possible values at a single point in the JSON tree. the AND condition across multiple target points is already served by using multiple values. This implementation should allow both AND conditions in different positions and OR conditions within a position at the same time.